+++ src/store/auth-context.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const AuthContext = React.createContext({
+  isLogin: false,
+  onLogin: (email, password) => {},
+});
+
+export default AuthContext;
